Ads hash table data structure.
The hashing function was swiped from Perl 5.005. The hash table itself
is about as standard as they come. Colliding hashes chain into a
doubly-linked list (for no other reason that its code was handy) that
records both keys and values.

The satellite data is a simple int, however the key-value list could be
easily modified to accept other kinds of data.