Ads an upper bound on linear probing.
The open adressing table now grows by powers of two, this allows us to
compute the `slack' (the threshold of how many consecutive probes the
table can do before its time to rezise) by simply adding 1 on each
resize.

This slack is also added as extra cells in the table. For example, a
table of size 32 will have a slack of 5. By adding these extra cells, we
can `overflow' while probing, without having to compute a modulus
operation for the wrap-around.

This is safe, because the moment we would try to probe outside the
bounds of the table, it would immediately rezise, as we would have
probes beyond the slack threshold.